+++ CHANGELOG.md
@@ -5,11 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2024-03-10
+
+### Added
+
+- Image comparison tools
+
+#### Features
+
+- Basic Visual testing API for Page object
+
 ## [0.2.5] - 2024-03-05
 
 ### Added
 
 - 'windows application driver' automation name, allowing directly use WinApplicationDriver without appium
+
+#### Automation Tools support
+
+- Window Application Driver
 
 ## [0.2.4] - 2024-02-26